asterisk/etc/asterisk/services/requirements.txt
git+https://github.com/digium/swagger-py.git
ari
odoorpc
py-Asterisk

#next dependencies are installed with aline's apk but left here for reference
gevent
requests
setproctitle
